requirements.txt
moviepy==1.0.3
faster-whisper==1.0.2
openai==1.54.3
requests==2.32.3
python-dotenv==1.0.1
streamlit==1.39.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.5
scikit-learn==1.5.2
scipy==1.13.1
librosa==0.10.2
pydub==0.25.1
pytube==15.0.0
google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pysrt==1.1.2
webvtt-py==0.4.6
colorama==0.4.6
scenedetect==0.6.2
imageio==2.34.1
imageio-ffmpeg==0.5.1
matplotlib==3.9.0
seaborn==0.13.2
pandas==2.2.2
redis==5.0.7
pickle5==0.0.12
pyyaml==6.0.2
jinja2==3.1.4

